Adding an experiment to respect file extensions and allow Markdown.
When using this experiment .html files will no longer be assumed to
be Nunjucks files, they'll be treated as HTML.  This allows other
extensions like .md to also be respected so that's included in the
same experiment.  Knowing that plain Markdown output isn't right for
everyone Plugins can now use the experimental integration (which is
likely to change from this initial version) to provide Markdown
renderers to the users of the plugin.